fix: style flow fix for react, next, vue, rn, remix

+++ packages/codegen-cli/src/scripts/setup-tailwind.ts
@@ -1,37 +1,17 @@
 import fs from "fs-extra";
 import { logger } from "../utils/logger";
 import { execa } from "execa";
-import { prependLine, writeFiles } from "../utils/io-util";
+import {
+  addPackageToPackageJson,
+  prependLine,
+  writeFiles,
+} from "../utils/io-util";
 import * as styleTemplate from "../utils/style-template";
 
 type FileDetails = {
   path?: string;
   fileName: string;
   content: string;
-};
-
-type TailwindConfigTemplates = {
-  tailwindConfigTemplate: string;
-  tailwindFileName: string;
-  postcssConfigTemplate: string;
-  postcssFileName: string;
-};
-
-type Variants = "ts" | "js";
-
-const tailwindConfigMap: Record<Variants, TailwindConfigTemplates> = {
-  js: {
-    tailwindConfigTemplate: styleTemplate.TAILWIND_CONFIG_NEXT_JS,
-    tailwindFileName: "tailwind.config.js",
-    postcssConfigTemplate: styleTemplate.POSTCSS_CONFIG_NEXT_JS,
-    postcssFileName: "postcss.config.js",
-  },
-  ts: {
-    tailwindConfigTemplate: styleTemplate.TAILWIND_CONFIG_NEXT_TS,
-    tailwindFileName: "tailwind.config.ts",
-    postcssConfigTemplate: styleTemplate.POSTCSS_CONFIG_NEXT_MJS,
-    postcssFileName: "postcss.config.mjs",
-  },
 };
 
 const runFileOperationPipeLine = async (
@@ -49,7 +29,9 @@ const runFileOperationPipeLine = async (
   }
 
   for (const fileToRemove of fileToRemoves) {
-    await fs.remove(`${root}${fileToRemove}`);
+    if (fs.existsSync(`${root}${fileToRemove}`)) {
+      await fs.remove(`${root}${fileToRemove}`);
+    }
   }
 };
 
@@ -61,12 +43,13 @@ export async function setTailwindOnNextjs(rootDir: string, variant: string) {
     tailwindFileName,
     postcssConfigTemplate,
     postcssFileName,
-  } = tailwindConfigMap[variant as Variants];
+  } =
+    styleTemplate.tailwindConfigMap["next"][variant as styleTemplate.Variants];
 
   const fileOperations: FileDetails[] = [
     {
-      fileName: postcssFileName,
-      content: postcssConfigTemplate,
+      fileName: postcssFileName!,
+      content: postcssConfigTemplate!,
     },
     {
       fileName: tailwindFileName,
@@ -84,23 +67,26 @@ export async function setTailwindOnNextjs(rootDir: string, variant: string) {
     },
   ];
 
-  const fileToRemove = ["/app/page.module.css"];
+  const fileToRemove = ["/app/page.module.css", "/app/page.js"];
 
   try {
-    runFileOperationPipeLine(rootDir, fileOperations, fileToRemove);
+    await runFileOperationPipeLine(rootDir, fileOperations, fileToRemove);
   } catch (error) {
     logger.error(error);
   }
 }
 
 export async function setTailwindOnReact(rootDir: string, variant: string) {
+  logger.info(
+    "Setting up tailwind on React. This may take a few minutes depending on your system.",
+  );
   const { tailwindFileName, postcssConfigTemplate, postcssFileName } =
-    tailwindConfigMap[variant as Variants];
+    styleTemplate.tailwindConfigMap["react"][variant as styleTemplate.Variants];
 
   const fileOperations: FileDetails[] = [
     {
-      fileName: postcssFileName,
-      content: postcssConfigTemplate,
+      fileName: postcssFileName!,
+      content: postcssConfigTemplate!,
     },
     {
       fileName: tailwindFileName,
@@ -120,54 +106,48 @@ export async function setTailwindOnReact(rootDir: string, variant: string) {
 
   const fileToRemove = ["/src/App.css"];
   try {
-    runFileOperationPipeLine(rootDir, fileOperations, fileToRemove);
+    await runFileOperationPipeLine(rootDir, fileOperations, fileToRemove);
   } catch (error) {
     logger.error(error);
   }
 }
 
 export async function setTailwindOnVue(rootDir: string, variant?: string) {
-  try {
-    if (variant === "ts") {
-      logger.info("Setting up tailwind on vue with typescript in ", rootDir);
-      await execa("npx", ["tailwindcss", "init", "--ts", "-p"], {
-        cwd: rootDir,
-      });
+  const {
+    tailwindConfigTemplate,
+    tailwindFileName,
+    postcssConfigTemplate,
+    postcssFileName,
+  } = styleTemplate.tailwindConfigMap["vue"][variant as styleTemplate.Variants];
 
-      // await writeFiles({
-      //   root: rootDir,
-      //   fileName: "tailwind.config.ts",
-      //   content: styleTemplate.TAILWIND_CONFIG_VUE_TS,
-      // });
-    } else {
-      await execa("npx", ["tailwindcss", "init", "-p"], {
-        cwd: rootDir,
-      });
-
-      await writeFiles({
-        root: rootDir,
-        fileName: "tailwind.config.js",
-        content: styleTemplate.TAILWIND_CONFIG_VUE_JS,
-      });
-    }
-
-    await writeFiles({
-      root: `${rootDir}/src`,
+  const fileOperations: FileDetails[] = [
+    {
+      fileName: postcssFileName!,
+      content: postcssConfigTemplate!,
+    },
+    {
+      fileName: tailwindFileName,
+      content: tailwindConfigTemplate,
+    },
+    {
+      path: "/src",
       fileName: "style.css",
       content: styleTemplate.TAILWIND_GLOBAL_CSS,
-    });
-
-    await writeFiles({
-      root: `${rootDir}/src`,
+    },
+    {
+      path: "/src",
       fileName: "App.vue",
       content: styleTemplate.VUE_APP_PAGE,
-    });
-
-    await writeFiles({
-      root: `${rootDir}/src/components`,
+    },
+    {
+      path: "/src/components",
       fileName: "HelloWorld.vue",
       content: styleTemplate.VUE_HELLOWORLD_PAGE,
-    });
+    },
+  ];
+
+  try {
+    await runFileOperationPipeLine(rootDir, fileOperations, []);
   } catch (error) {
     logger.error(error);
   }
@@ -178,6 +158,10 @@ export async function setNativewindOnExpo(rootDir: string) {
     /* TODO: Update Nativewind as of now using nativewind v2.x and tailwindcss v3.3.2
      *  as nativewind v4.x get into stable we will use tailwindcss@latest with expo.
      */
+
+    await addPackageToPackageJson(rootDir, {
+      devDependencies: ["nativewind", "tailwindcss"],
+    });
 
     // if OS is mac run pod install
     if (process.platform === "darwin") {
@@ -190,7 +174,25 @@ export async function setNativewindOnExpo(rootDir: string) {
       content: styleTemplate.NATIVEWIND_TAILWIND_CONFIG,
     });
 
-    await prependLine(`${rootDir}/global.css`, 'import "../global.css"');
+    await writeFiles({
+      root: rootDir,
+      fileName: "global.css",
+      content: styleTemplate.TAILWIND_GLOBAL_CSS,
+    });
+
+    await writeFiles({
+      root: rootDir,
+      fileName: "babel.config.js",
+      content: styleTemplate.NATIVEWIND_BABEL_CONFIG_jS,
+    });
+
+    await writeFiles({
+      root: rootDir,
+      fileName: "metro.config.js",
+      content: styleTemplate.NATIVEWIND_METRO_CONFIG_JS,
+    });
+
+    await prependLine(`${rootDir}/app/_layout.tsx`, 'import "../global.css"');
 
     await writeFiles({
       root: rootDir,
